Calculate deep percolation differently
We change deep percolation so that it uses theta-theta_fc rather than
theta_s-theta_fc. We also take rainfall into account. This makes the
model be in line with @nmalamos's Excel files.

However this is not documented (the documentation still says
theta_s-theta_fc). In addition, the term "draintime" is now probably
wrong. We will fix these later.